+++ src/bract/cli/inducer.clj
@@ -13,6 +13,7 @@
     [clojure.pprint     :as pp]
     [clojure.tools.cli  :as cli]
     [keypin.util        :as kputil]
+    [bract.core.echo    :as echo]
     [bract.core.keydef  :as core-kdef]
     [bract.core.inducer :as core-inducer]
     [bract.core.util    :as core-util]
@@ -21,12 +22,34 @@
 
 
 (defn merge-commands
-  "Merge given CLI command map (key: command-string, value: {:doc string :handler (fn [context]) -> context}) to the
-  existing one."
+  "Merge given CLI command map {\"command-name\" {:doc string :handler inducer}} to the existing one. The inducer is
+  `(fn [context]) -> context`."
   [context new-commands]
-  (let [old-commands (clim-kdef/ctx-app-commands context)]
-    (assoc context
-      (key clim-kdef/ctx-app-commands) (merge old-commands new-commands))))
+  (echo/echo "Adding CLI commands:" (keys new-commands))
+  (->> new-commands
+    (merge (clim-kdef/ctx-app-commands context))  ; merge/assoc instead of update to enforce app-commands validation
+    (assoc context (key clim-kdef/ctx-app-commands))))
+
+
+(defn merge-launch-commands
+  "Merge given CLI command map {\"command-name\" {:doc string :handler launcher}} to the existing one. The launcher is
+  `(fn [context]) -> context` meant to replace the entry at key :bract.core/launcher in the context."
+  [context new-commands]
+  (echo/echo "Adding CLI commands for launchers:" (->> (vals new-commands)
+                                                    (map :handler)
+                                                    (zipmap (keys new-commands))))
+  (->> new-commands
+    (reduce-kv (fn [m k command]
+                 (->> (fn [launcher]
+                        (fn update-launcher [context]
+                          (assoc context
+                            (key core-kdef/ctx-launch?) true
+                            (key core-kdef/ctx-launcher) launcher)))
+                   (update command :handler)
+                   (assoc m k)))
+      {})
+    (merge (clim-kdef/ctx-app-commands context))  ; merge/assoc instead of update to enforce app-commands validation
+    (assoc context (key clim-kdef/ctx-app-commands))))
 
 
 (defn parse-args